Refactor rule and decoder tool into utils module
